donateX/app/requirements.txt
alembic==1.8.1
CouchDB==1.2
fastapi==0.100.0
mysql-connector-python==8.0.32
pydantic==2.1.1
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2022.7.1
singleton==0.1.0
SQLAlchemy==1.4.46
starlette==0.27.0
typing_extensions==4.7.1
uvicorn==0.20.0
websockets==10.4
sse-starlette==1.6.1
pyfcm==1.5.4
pandas==2.0.3
numpy==1.26.4
openpyxl==3.1.2
python-multipart==0.0.6
cryptography~=40.0.2
Jinja2~=3.1.2
google-auth~=2.21.0
google-api-core==2.11.1
google-api-python-client==2.92.0
google-auth==2.21.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.62.0
python-jose[cryptography]
paypalrestsdk==1.13.3
streamlit==1.44.1